+++ src/components/idConfigSelectRow.js
@@ -1,6 +1,6 @@
 const { StringSelectMenuBuilder, ActionRowBuilder } = require("discord.js");
 
-const buildIdConfigSelectRow = (ids, config, placeholder, data, eventName, showId=true) => {
+const buildIdConfigSelectRow = (ids, config, placeholder, data, eventName, showId=true, extraFields=[]) => {
     const menuId = {
         eventName: eventName,
         ...data,
@@ -11,8 +11,11 @@ const buildIdConfigSelectRow = (ids, config, placeholder, data, eventName, showI
         .setPlaceholder(placeholder)
         .addOptions(ids.map(id => {
             const data = config[id];
+            const extra = extraFields.map(field => {
+                return data[field];
+            }).join(" | ");
             return {
-                label: `${data.name}${showId ? " #" + id.toString() : ""}`,
+                label: `${data.name}${extra ? " | " + extra : ""}${showId ? " #" + id.toString() : ""}`,
                 value: `${id}`,
             }
         }));