+++ src/config/pokemonConfig.js
@@ -2456,6 +2456,44 @@ const pokemonConfig = {
         "rarity": rarities.LEGENDARY,
         "growthRate": growthRates.SLOW
     },
+    "151": {
+        "name": "Mew",
+        "emoji": "<:151:1116755839919853670>",
+        "description": "So rare that it is still said to be a mirage by many experts. Only a few people have seen it worldwide.",
+        "type": [types.PSYCHIC],
+        "baseStats": [100, 100, 100, 100, 100, 100],
+        "sprite": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/151.png",
+        "shinySprite": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/shiny/151.png",
+        "abilities": {
+            "28": 1
+        },
+        "moveIds": ["m246", "m94", "m396", "m63"],
+        "battleEligible": true,
+        "rarity": rarities.MYTHICAL,
+        "growthRate": growthRates.MEDIUMSLOW,
+        "noGacha": true,
+        "mythicConfig": {
+            "basicMoveIds": [
+                "m16", "m17", "m22", "m35", "m40", "m43", "m51", "m52", "m55", "m71", "m77", "m81", 
+                "m84", "m93", "m98", "m100", "m102", "m106", "m116", "m118", "m122", "m123", "m175", 
+                "m189", "m205", "m214", "m239", "m246", "m270", "m418", "m420", "m479", "m506",
+                "m574", "m876"
+            ],
+            "powerMoveIds": [
+                "m14", "m34", "m36", "m46",  "m53", "m57", "m58", "m60", "m65", "m68",
+                "m70", "m85", "m86", "m91", "m92", "m94", "m113", "m115", "m127", "m182", "m188", "m191",
+                "m203", "m215",  "m216", "m219", "m226", "m247", "m269", "m202", "m276", "m283",
+                "m317", "m334", "m340", "m352", "m355", "m369", "m387", "m396", "m398", "m402",
+                "m409", "m414", "m417", "m424", "m430", "m441", "m444", "m446", "m450", "m521", "m523",
+                "m525", "m529"
+            ],
+            "ultimateMoveIds": [
+                "m38", "m56", "m63", "m76", "m87", "m89", "m135", "m143", "m153", "m156", "m157", 
+                "m162", "m200", "m223", "m257", "m304", "m366", "m394", "m405", "m413", "m416", 
+                "m428", "m433", "m437", "m482", "m528", "m542"
+            ]
+        }
+    },
     "20091": {
         "name": "Temple Guardian Cloyster",
         "emoji": "<:91:1100288966881718342>",
@@ -2604,7 +2642,7 @@ const rarityConfig = {
     [rarities.MYTHICAL]: {
         pokemon: [],
         color: 0xff0000,
-        money: 1000,
+        money: 10000,
     },
 }